Restructure match highlighting to prevent XSS
The previous code generated inline HTML and relied on ng-bind-html being
used, which created an attack vector.

The new approach divides the string into three parts: before the match,
the match, and after the match. These can then be looped over in the
template, and the match class applied appropriately, while benefitting
from Angular's automatic sanitization of the content.

Once matchAll gets some traction it should be possible to apply this to
multiple matches as well.